Introduced the concept of a dedicated kernel stack for each thread, refactored context.S intr.S and vectors.S for this

+++ kernel/platform/malta/syscall.c
@@ -6,7 +6,7 @@
 #include <sys/cdefs.h>
 #include "interrupts.h"
 
-int syscall(int number, ...)
+long int syscall(long int number, ...)
 {
     // NOTE: apparently the vaargs things needs a real stack frame to work (??)
     __unused volatile char tmp[8] = {0};
